SLA central data changed to float

+++ src/main/java/com/synectiks/asset/response/DataProtectionResponse.java
@@ -17,6 +17,6 @@ public class DataProtectionResponse implements Serializable {
 
   private static final long serialVersionUID = 1L;
 
-  private Integer score;
+  private Float score;
   
 }
